[Feat/#290] :wrench: Refactor: 제출 시간 받도록 수정
