+++ chatty-be/src/main/java/com/chatty/chatty/game/controller/dto/SubmitAnswerRequest.java
@@ -1,10 +1,14 @@
 package com.chatty.chatty.game.controller.dto;
 
+import java.time.LocalDateTime;
+
 public record SubmitAnswerRequest(
         Integer quizNum,
 
         Long playerId,
 
-        String playerAnswer
+        String playerAnswer,
+
+        LocalDateTime submittedTime
 ) {
 }
