+++ chatty-be/src/main/java/com/chatty/chatty/game/controller/dto/model/AnswerDTO.java
@@ -1,9 +1,9 @@
-package com.chatty.chatty.game.domain;
+package com.chatty.chatty.game.controller.dto.model;
 
 import lombok.Builder;
 
 @Builder
-public record Answer(
+public record AnswerDTO(
         Long user_id,
         String user
 ) {
