+++ chatty-be/src/main/java/com/chatty/chatty/game/domain/AnswerData.java
@@ -1,24 +1,25 @@
 package com.chatty.chatty.game.domain;
 
-import java.util.ArrayList;
-import java.util.List;
+import com.chatty.chatty.game.controller.dto.SubmitAnswerRequest;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 import lombok.Builder;
 import lombok.Getter;
 
 @Builder
 @Getter
 public class AnswerData {
-    private final List<Answer> playerAnswers = new ArrayList<>();
+    private final Map<Long, PlayerAnswerData> playerAnswers = new ConcurrentHashMap<>();
     private final Integer playerNum;
     private final Integer majorityNum;
     private final String quizId;
     private final Integer quizNum;
     private final String correct;
 
-    public synchronized SubmitStatus addAnswer(Long playerId, String playerAnswer) {
-        playerAnswers.add(Answer.builder()
-                .user_id(playerId)
-                .user(playerAnswer)
+    public synchronized SubmitStatus addAnswer(SubmitAnswerRequest request) {
+        playerAnswers.put(request.playerId(), PlayerAnswerData.builder()
+                .playerAnswer(request.playerAnswer())
+                .submittedTime(request.submittedTime())
                 .build());
         return checkSubmitStatus();
     }
